+++ packages/form/src/tip.vue
@@ -1,7 +1,11 @@
 <template>
   <div class="weui-form__tips-area">
-    <p class="weui-form__tips" v-if="tipContent">{{ tipContent }}</p>
-    <p class="weui-form__tips" v-else-if="content">{{ content }}</p>
+    <p class="weui-form__tips">
+      <template v-if="tipContent">{{ tipContent }}</template>
+      <template v-else>
+        <slot name="tipContent"></slot>
+      </template>
+    </p>
     <slot name="content"></slot>
   </div>
 </template>