+++ types/component.d.ts
@@ -1,0 +1,7 @@
+import Vue from 'vue'
+
+export class WeuiComponent extends Vue {
+    static name: string
+
+    static install(vue: typeof Vue): void
+}